+++ CHANGELOG.md
@@ -15,12 +15,16 @@ Note that since we don't clearly distinguish between a public and private interf
 
 ## [Unreleased]
 
+- `meshes` extension: Fixed a bug in mesh visualization (show backfaces when opacity < 1) 
+- Add color quick select control to Volume controls
+
 ## [v3.28.0] - 2022-12-20
 
 - Show histogram in direct volume control point settings
 - Add `solidInterior` parameter to sphere/cylinder impostors
 - [Breaking] Tweak `ignoreHydrogens` non-polar handling (introduced in 3.27.0)
 - Add `meshes` and `volumes-and-segmentations` extensions
+    - See https://molstarvolseg.ncbr.muni.cz/ for more info
 - Fix missing support for info in `ParamDefinition.Converted`
 - Add support for multi-visual volume representations
 - Improve volume isosurface bounding-sphere